The delay was calculated wrong when the averageDT was used. Before it was dist*speed, which would give a property in pixel^2/s... We want the delay to be a time in us, so now I do dist/speed, but importantly before speed is converted to PPS so that we get us and not seconds. Also the calculation of delay is more accurate if the dist is used as a float instead of an int.
Importantly, delay was of type short before. However the delay can easily get larger than 32ms, so short is not sufficient. I changed the type in MotionOrientationEvent to int.

git-svn-id: https://svn.code.sf.net/p/jaer/code/branches/jAER_1.5@4694 b7f4320f-462c-0410-a916-d9f35bb82d52

+++ src/net/sf/jaer/event/MotionOrientationEvent.java
@@ -20,10 +20,10 @@ public class MotionOrientationEvent extends DvsOrientationEvent {
     /** unit vector of direction of motion */
     public Dir dir=null;
     
-    /** the 'delay' value of this cell, an analog (but quantized) quantity 
-     * that signals the time delay associated with this event.
+    /** the 'delay' value of this cell in us (unit of timestamps), an analog 
+     * (but quantized) quantity that signals the time delay associated with this event.
      * Smaller time delays signal larger speed */
-    public short delay=0;
+    public int delay=0;
     
     /** the distance associated with this motion event. This is the 
      * distance in pixels to the prior event that signaled this direction */
@@ -85,7 +85,8 @@ public class MotionOrientationEvent extends DvsOrientationEvent {
     static public Point2D.Float computeMotionVector(MotionOrientationEvent e){
         Dir d=unitDirs[e.direction];
         int dist=e.distance;
-        int delay=e.delay; if(delay==0) delay=1;
+        int delay=e.delay; 
+        if(delay==0) delay=1;
         float speed=( (float)dist / ((float)delay*1e-6f) );
         motionVector.setLocation(d.x*speed,d.y*speed);
         return motionVector;
